+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 NOTE: Subcrates have their own changelogs: [bevy-tnua-physics-integration-layer](physics-integration-layer/CHANGELOG.md), [bevy-tnua-rapier](rapier3d/CHANGELOG.md), [bevy-tnua-xpbd](xpbd3d/CHANGELOG.md).
 
 ## [Unreleased]
+## 0.17.0 - 2024-05-07
 ### Removed
 - [**BREAKING**] `TnuaBuiltinWalk` no longer has an `up` field. The up
   direction is fixed to `Direction3d::Y` (up until now, it problably wouln't